refactored (un)signed byte/short packet read (java uses signed data types everywhere, readSC and readSH were complete nonsense)

+++ chat-server/src/com/aionemu/chatserver/network/gameserver/clientpackets/CM_CS_AUTH.java
@@ -34,7 +34,7 @@ public class CM_CS_AUTH extends GsClientPacket {
 
 	@Override
 	protected void readImpl() {
-		gameServerId = (byte) readC();
+		gameServerId = readC();
 		password = readS();
 	}
