+++ game-server/src/com/aionemu/gameserver/network/aion/clientpackets/CM_FRIEND_STATUS.java
@@ -27,7 +27,7 @@ public class CM_FRIEND_STATUS extends AionClientPacket {
 
 	@Override
 	protected void readImpl() {
-		status = (byte) readC();
+		status = readC();
 	}
 
 	@Override
