+++ login-server/src/com/aionemu/loginserver/network/gameserver/clientpackets/CM_GS_PONG.java
@@ -13,7 +13,7 @@ public class CM_GS_PONG extends GsClientPacket {
 
 	@Override
 	protected void readImpl() {
-		serverId = (byte) readC();
+		serverId = readC();
 		pid = readD();
 	}
