+++ login-server/src/com/aionemu/loginserver/network/gameserver/clientpackets/CM_PREMIUM_CONTROL.java
@@ -18,7 +18,7 @@ public class CM_PREMIUM_CONTROL extends GsClientPacket {
 		accountId = readD();
 		requestId = readD();
 		requiredCost = readQ();
-		serverId = (byte) readC();
+		serverId = readC();
 	}
 
 	@Override
